Add initial GUI Remove 'playGame' logic
Add initial GUI that enables the user to interact and play the game
through the web page.

Remove the 'playGame' logic that was used to start the game through the
console.
Now that the initial GUI was created, the 'playGame' logic was removed
from the code as it was no longer needed.